Changes, bug fixes and improvements.
- i'd like to think at least.

so many memory leaks and bugs out there to be squashed, one day i'll
figure the right type of memory management scheme out and implement it.

- implemented ObjectArray

- changed and removed all references to ReturnResult and swapped them
for Object* references. Code changed to make that happen. It isn't ideal
as I just got it working from the old code base. Needs refining soon,
because of redundant use of Object and Object*'s among other things.

- because of this, many results and underlying types have been moved
around and worked wtih.

AUTO - GENERATED files have been hand coded, so the code generators need
updating before we regerenate code and replace hand written things.

Other stuff i'm too tired to mention.

+++ src/Lox.h
@@ -38,8 +38,8 @@ struct _Loxfuncs{
 
 struct _Lox {
 
-	int hadError;
-     int hadRuntimeError;
+	short int hadError;
+    short int hadRuntimeError;
 	Interpreter interpreter;
 	void (*run)(Lox* lox,char*);
 	void (*runPrompt)(Lox* lox);
